feature : 스터디룸 생성, 목록 보기, 입장 로직 구현

+++ src/main/java/GRWM/backend/dto/studyroom/StudyRoomListDto.java
@@ -11,7 +11,7 @@ import java.util.List;
 @NoArgsConstructor
 public class StudyRoomListDto {
     List<StudyRoomBriefDto> studyRooms;
-    int totalElement; // 스터디룸의 총 개수
+    Long totalElement; // 스터디룸의 총 개수
     int currentPage; // 페이징 시 현재 보내는 페이지
     int totalPages; // 전체 페이지 개수 (페이징 관련 파라미터)
